backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
python-dotenv==1.0.0
google-generativeai==0.3.1
httpx==0.24.1
python-multipart==0.0.6
jinja2==3.1.2 